Greatly improved retransmit requesting. Retransmit requests are now re-sent after 100ms if they are unanswered

+++ src/main/java/org/phlo/AirReceiver/RaopRtpPacket.java
@@ -201,7 +201,7 @@ public abstract class RaopRtpPacket extends RtpPacket {
 			setBeUInt16(getBuffer(), RaopRtpPacket.Length, value);
 		}
 
-		public long getSequenceCount() {
+		public int getSequenceCount() {
 			return getBeUInt16(getBuffer(), RaopRtpPacket.Length + 2);
 		}
